backend/app/services/llm_service.py
"""
LLM Service for OpenAI and Anthropic integration
"""
from typing import Optional, Dict, Any
import logging
from langchain_openai import ChatOpenAI
from langchain_anthropic import ChatAnthropic
from langchain_core.messages import HumanMessage, SystemMessage

from app.config import settings

logger = logging.getLogger(__name__)


class LLMService:
    """Service for interacting with Large Language Models"""
    
    def __init__(self):
        self.provider = settings.DEFAULT_LLM_PROVIDER
        self.model_name = settings.DEFAULT_MODEL
        self._llm = None
    
    @property
    def llm(self):
        """Lazy load LLM instance"""
        if self._llm is None:
            self._llm = self._create_llm()
        return self._llm
    
    def _create_llm(self):
        """Create LLM instance based on provider"""
        if self.provider == "openai":
            if not settings.OPENAI_API_KEY:
                raise ValueError("OPENAI_API_KEY not configured")
            return ChatOpenAI(
                model=self.model_name,
                temperature=0,
                api_key=settings.OPENAI_API_KEY
            )
        elif self.provider == "anthropic":
            if not settings.ANTHROPIC_API_KEY:
                raise ValueError("ANTHROPIC_API_KEY not configured")
            return ChatAnthropic(
                model=self.model_name,
                temperature=0,
                api_key=settings.ANTHROPIC_API_KEY
            )
        else:
            raise ValueError(f"Unsupported LLM provider: {self.provider}")
    
    async def classify_document(self, text: str) -> Dict[str, Any]:
        """Classify document type and extract metadata"""
        # TODO: Implement LLM-based classification
        pass
    
    async def extract_renewal_dates(self, text: str) -> list:
        """Extract renewal dates from text"""
        # TODO: Implement LLM-based extraction
        pass
    
    async def extract_obligations(self, text: str, document_type: str) -> list:
        """Extract contractual obligations"""
        # TODO: Implement LLM-based extraction
        pass
    
    async def extract_compliance_requirements(self, text: str) -> list:
        """Extract compliance requirements"""
        # TODO: Implement LLM-based extraction
        pass


# Global instance
llm_service = LLMService()
