backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# LangChain & LangGraph
langchain
langchain-openai
langchain-anthropic
langchain-community
langgraph
langgraph-checkpoint

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
pillow==10.4.0

# OCR (optional)
pytesseract==0.3.13

# NLP
# spacy==3.8.2  # Commented out - requires C compiler
python-dateutil==2.9.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
redis==5.2.0

# Vector Store
# chromadb==0.5.20  # Commented out - optional for now

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# Logging & Monitoring
structlog==24.4.0
